Moved halfbinary to correct package
